Use manager per queue in tensor implementation
Move implementation from dpctl/utils/__init__.py into
dedicated files.
